fix(ci): wait for linux binaries when exporting to web
Ensures we have both the linux and wasm binaries when exporting to web.
The linux binaries are required as the host system uses linux and will
fail to identify the GDExtension node types (PTY, Terminal) leading to
the export not including them.

Also removes the retry action as the initial export command was not
returning an error exit code anyway.