fix: prevent visual glitch when resizing terminal
Hides fore_canvas_item when doing a full redraw (e.g. on resize) to
prevent showing a corrupted texture from the viewport when it is
resized.

When the fore_canvas_item is hidden, draws directly to the main canvas
item. This requires cell attributes, which are typically handled by the
shader (blink, inverse video), to be handled by C++ code.

For now, blinking characters are always made visible as it is difficult
to sync with the shaders TIME value. Also, it could be considered good
UX to show characters that might be hidden when the user is resizing.