fix(input): handle AltGr combinations
On Windows, some international keyboards use Ctrl+Alt (AltGr)
to type special chars (e.g ']' with Ctrl+Alt+9). Previously,
both the unicode character AND raw modifiers were sent,
causing "^<char>" instead of "<char>" ('^]' instead of ']'
in the example).

This PR handles the special case of ctrl+alt on Windows so that when
ctrl+alt is pressed, there is a unicode character, and the key/unicode
pair cannot be found in the KEY_MAP (e.g {KEY_9, ']'}), then we assume
the ctrl+alt modifiers where used to produce the unicode character
and strip them.

It also fixes some test cases (uses the unicode value so it is only 0 on
Windows). And adds some test cases to ensure backwards compatability for
Linux and compatability with the way Linux handles AltGr (i.e. does not
set ctrl+alt, allowing for both AltGr+9 = ']' and Ctrl+AltGr+9 = '^]').